+++ packages/eas-cli/src/commandUtils/context/ContextField.ts
@@ -1,5 +1,4 @@
 import { Analytics } from '../../analytics/AnalyticsManager';
-import { EnvironmentVariableEnvironment } from '../../build/utils/environment';
 import SessionManager from '../../user/SessionManager';
 import { Client } from '../../vcs/vcs';
 
@@ -11,7 +10,7 @@ export interface ContextOptions {
   /**
    * If specified, env variables from the selected environment will be fetched from the server and used to evaluate the dynamic config.
    */
-  withServerSideEnvironment?: EnvironmentVariableEnvironment | null;
+  withServerSideEnvironment?: string | null;
 }
 
 export default abstract class ContextField<T> {