+++ packages/eas-cli/src/graphql/queries/EnvironmentVariablesQuery.ts
@@ -1,7 +1,6 @@
 import { print } from 'graphql';
 import gql from 'graphql-tag';
 
-import { EnvironmentVariableEnvironment } from '../../build/utils/environment';
 import { ExpoGraphqlClient } from '../../commandUtils/context/contextUtils/createGraphqlClient';
 import { withErrorHandlingAsync } from '../client';
 import {
@@ -14,7 +13,7 @@ import { EnvironmentVariableFragmentNode } from '../types/EnvironmentVariable';
 import { EnvironmentVariableWithSecretFragmentNode } from '../types/EnvironmentVariableWithSecret';
 
 type EnvironmentVariableWithLinkedEnvironments = EnvironmentVariableFragment & {
-  linkedEnvironments?: EnvironmentVariableEnvironment[] | null;
+  linkedEnvironments?: string[] | null;
 };
 
 export type EnvironmentVariableWithFileContent = EnvironmentVariableFragment & {
@@ -22,6 +21,31 @@ export type EnvironmentVariableWithFileContent = EnvironmentVariableFragment & {
 };
 
 export const EnvironmentVariablesQuery = {
+  async environmentVariableEnvironmentsAsync(
+    graphqlClient: ExpoGraphqlClient,
+    appId: string
+  ): Promise<string[]> {
+    const data = await withErrorHandlingAsync(
+      graphqlClient
+        .query(
+          gql`
+            query AppEnvironmentVariableEnvironments($appId: String!) {
+              app {
+                byId(appId: $appId) {
+                  id
+                  environmentVariableEnvironments
+                }
+              }
+            }
+          `,
+          { appId },
+          { additionalTypenames: ['App'] }
+        )
+        .toPromise()
+    );
+
+    return data.app?.byId.environmentVariableEnvironments ?? [];
+  },
   async byAppIdWithSensitiveAsync(
     graphqlClient: ExpoGraphqlClient,
     {
@@ -31,7 +55,7 @@ export const EnvironmentVariablesQuery = {
       includeFileContent = false,
     }: {
       appId: string;
-      environment?: EnvironmentVariableEnvironment;
+      environment?: string;
       filterNames?: string[];
       includeFileContent?: boolean;
     }
@@ -80,7 +104,7 @@ export const EnvironmentVariablesQuery = {
       includeFileContent = false,
     }: {
       appId: string;
-      environment?: EnvironmentVariableEnvironment;
+      environment?: string;
       filterNames?: string[];
       includeFileContent?: boolean;
     }
@@ -128,7 +152,7 @@ export const EnvironmentVariablesQuery = {
     }: {
       appId: string;
       filterNames?: string[];
-      environment?: EnvironmentVariableEnvironment;
+      environment?: string;
       includeFileContent?: boolean;
     }
   ): Promise<(EnvironmentVariableWithFileContent & EnvironmentVariableWithLinkedEnvironments)[]> {
@@ -178,7 +202,7 @@ export const EnvironmentVariablesQuery = {
     }: {
       appId: string;
       filterNames?: string[];
-      environment?: EnvironmentVariableEnvironment;
+      environment?: string;
       includeFileContent?: boolean;
     }
   ): Promise<EnvironmentVariableWithFileContent[]> {